created abstracted Service class to avoid repetative boilerplate code
