+++ pi_dashboard/app/http/service.py
@@ -1,0 +1,10 @@
+import os
+
+class Service(object):
+  def __init__(self):
+    # find token and setup headers object
+    token = os.getenv('WANIKANI_TOKEN')
+    if not token:
+      raise Exception('secret token not found')
+
+    self.headers = {'Authorization': 'Bearer ' + token}
